requirements.txt
duckdb==0.8.0
pandas==2.0.2
openai==0.27.7
python-dotenv==1.0.0
retry==0.9.2
flask
autoviz
pretty_html_table
